backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.10.0
    # via
    #   django
    #   django-cors-headers
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
certifi==2025.10.5
    # via requests
charset-normalizer==3.4.3
    # via requests
django==5.2.7
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   django-ical
    #   django-recurrence
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-cors-headers==4.9.0
    # via -r requirements.in
django-filter==25.2
    # via -r requirements.in
django-ical==1.9.2
    # via -r requirements.in
django-recurrence==1.11.1
    # via django-ical
django-stubs==5.2.7
    # via
    #   -r requirements.in
    #   djangorestframework-stubs
django-stubs-ext==5.2.7
    # via
    #   -r requirements.in
    #   django-stubs
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-camel-case==1.4.2
    # via -r requirements.in
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
djangorestframework-stubs==3.16.4
    # via -r requirements.in
drf-spectacular==0.28.0
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
icalendar==6.3.1
    # via
    #   -r requirements.in
    #   django-ical
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.25.1
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
packaging==25.0
    # via gunicorn
pillow==11.3.0
    # via -r requirements.in
psycopg2-binary==2.9.11
    # via -r requirements.in
pyjwt==2.10.1
    # via djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via
    #   django-recurrence
    #   icalendar
python-dotenv==1.1.1
    # via -r requirements.in
pyyaml==6.0.3
    # via drf-spectacular
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via djangorestframework-stubs
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
types-pyyaml==6.0.12.20250915
    # via
    #   django-stubs
    #   djangorestframework-stubs
types-requests==2.32.4.20250913
    # via djangorestframework-stubs
typing-extensions==4.15.0
    # via
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
tzdata==2025.2
    # via icalendar
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.5.0
    # via
    #   requests
    #   types-requests
